requirements.txt
fastapi
uvicorn
pydantic
mysql-connector-python
reportlab
python-multipart
psycopg2-binary
pytz
passlib
